TD-300 KANBAN: active my task chưa đúng và khi click  my task nó chưa không hiện ra các task của mình
